backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
backoff==2.2.1
bcrypt==4.3.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
dnspython==2.7.0
docker==7.1.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastapi-utils==0.8.0
fonttools==4.60.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
inference-cli==0.58.3
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
nvidia-ml-py==12.575.51
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
propcache==0.4.1
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyasn1==0.6.1
pybase64==1.0.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
requests==2.32.5
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
scipy==1.16.2
sentry-sdk==2.42.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.24
starlette==0.47.3
supervision==0.26.1
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
